+++ core/commands/node_capabilities_scan.go
@@ -6,30 +6,28 @@ import (
 )
 
 type (
-	CmdNodePrintCapabilities struct {
+	CmdNodeCapabilitiesScan struct {
 		OptsGlobal
 	}
 )
 
-func (t *CmdNodePrintCapabilities) Run() error {
+func (t *CmdNodeCapabilitiesScan) Run() error {
 	return nodeaction.New(
+		nodeaction.WithLocal(t.Local),
+		nodeaction.WithRemoteNodes(t.NodeSelector),
 		nodeaction.WithFormat(t.Format),
 		nodeaction.WithColor(t.Color),
 		nodeaction.WithServer(t.Server),
-
-		nodeaction.WithRemoteNodes(t.NodeSelector),
-		nodeaction.WithRemoteAction("node print capabilities"),
+		nodeaction.WithRemoteAction("scan capabilities"),
 		nodeaction.WithRemoteOptions(map[string]interface{}{
 			"format": t.Format,
 		}),
-
-		nodeaction.WithLocal(t.Local),
 		nodeaction.WithLocalRun(func() (interface{}, error) {
 			n, err := object.NewNode()
 			if err != nil {
 				return nil, err
 			}
-			return n.PrintCapabilities()
+			return n.ScanCapabilities()
 		}),
 	).Do()
 }